fix https://github.com/prisma/prisma-engines/issues/869
