+++ libs/datamodel/core/tests/config/mod.rs
@@ -1,3 +1,3 @@
-pub mod builtin;
+pub mod datasources_to_mcf;
 pub mod generators;
 pub mod sources;
